// src/services/admin/queue/interfaces/get-all-queue.types.ts
export interface IGetAllQueueResponse {
  appointments: Queue[];
  pagination: Pagination;
}

export interface Queue {
  id: string;
  queue_number: string;
  queue_date: string;
  description: string;
  status: string;
  patient: Patient;
}

export interface Patient {
  id: string;
  name: string;
}

export interface Pagination {
  current_page: number;
  first_page_url: string;
  from: number;
  last_page: number;
  last_page_url: string;
  links: Link[];
  next_page_url: string;
  path: string;
  per_page: number;
  prev_page_url: null;
  to: number;
  total: number;
}


export interface Link {
  url: null | string;
  label: string;
  active: boolean;
}